testing/mochitest/tests/Harness_sanity/file_SpecialPowersFrame1.html

<html>
  <head>
    <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  </head>
  <body>
    <div id="content" style="display: none">
      <script type="text/javascript">
        is(SpecialPowers.sanityCheck(), "foo", "Check Special Powers in iframe");
      </script>
    </div>
  </body>
</html>
